# hek_analysis/docs/architecture/Unified-Healthcare-API_openapi.yaml
openapi: 3.0.3
info:
  title: Unified Healthcare API
  description: >
    Canonical contract consolidating HISO, KARO, and ERMS. See
    Unified-Healthcare-API_API-Contract-Design.md for full rationale,
    traceability, and reconciliation decisions. No version identifier
    appears in the URL/header/media-type per stakeholder decision (Section 9
    of the contract document) — version is tracked here via info.version and
    an append-only changelog only; no request/response shape changes without
    a coordinated migration.
  version: "1.1.1"
  x-changelog:
    - version: "1.0.0"
      date: "2026-07-19"
      description: Initial canonical contract, Phase 7 (API Contract Design).
    - version: "1.1.0"
      date: "2026-07-19"
      description: >
        Demographics split into 4 legacy-shaped endpoints instead of one merged
        endpoint; HISO's addInvoice/launchForm actions and static-mode form
        retrieval implemented (previously excluded as dead stubs); KARO's
        GetEncounterSummary and SaveScreeningCode implemented for real
        (previously excluded as confirmed-broken); rate limiting confirmed
        as a built, config-toggled capability.
    - version: "1.1.1"
      date: "2026-07-20"
      description: >
        Block 3 contract re-verification against the actual Day-1 build (all 42
        operations implemented, route-for-route and verb-for-verb match).
        One drift corrected here rather than left silently mismatched:
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /patients/{patientId}/encounters/{encounterId}/notes:
    get:
      tags: [Clinical Notes]
      summary: Retrieve clinical/consult notes
      parameters:
        - $ref: '#/components/parameters/patientId'
        - $ref: '#/components/parameters/encounterId'
        - $ref: '#/components/parameters/sinceDate'
        - $ref: '#/components/parameters/untilDate'
        - $ref: '#/components/parameters/sortOrder'
      responses:
        '200':
          description: List of clinical notes (default lookback 24 months if no date range supplied)
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/ClinicalNote' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Clinical Notes]
      summary: Save a clinical/consult note
      parameters:
        - $ref: '#/components/parameters/patientId'
        - $ref: '#/components/parameters/encounterId'
        - $ref: '#/components/parameters/idempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClinicalNoteInput'
      responses:
        '201':
          description: Note created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ClinicalNote' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /patients/{patientId}/encounters/{encounterId}/conditions:
    get:
      tags: [Conditions]
      summary: Retrieve conditions/diagnoses
      parameters:
        - $ref: '#/components/parameters/patientId'
        - $ref: '#/components/parameters/encounterId'
      responses:
        '200':
          description: List of conditions
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Condition' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Conditions]
      summary: Save a condition/diagnosis (idempotent on duplicate)
      parameters:
        - $ref: '#/components/parameters/patientId'
        - $ref: '#/components/parameters/encounterId'
        - $ref: '#/components/parameters/idempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConditionInput'
      responses:
        '200':
          description: Duplicate submission — existing condition returned, not an error (FR-IDEM-01)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Condition' }
        '201':
          description: Condition created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Condition' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /acc45/sessions/{sessionKey}:
    get:
      tags: [ACC45]
      summary: Resolve session GUID to Provider/Patient/Appointment/Practice context
      parameters:
        - $ref: '#/components/parameters/sessionKey'
      responses:
        '200':
          description: Resolved session context
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SessionContext' }
        '401':
          description: Invalid session key (HISO-BR-02)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
  /acc45/sessions/{sessionKey}/forms:
    get:
      tags: [ACC45]
      summary: Retrieve form data via concept-mapping (HISO-BR-03/08/09/10)
      parameters:
        - $ref: '#/components/parameters/sessionKey'
        - name: mode
          in: query
          schema: { type: string, enum: [dynamic, static], default: dynamic }
          description: >
            Reflects the practice's IsDynamic setting. Legacy "static" mode was an
            empty, unimplemented stub (HISO-BR-17) — implemented for real per
            stakeholder decision 2026-07-19, not carried forward as a no-op.
      responses:
        '200':
          description: Form data container
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FormData' }
  /acc45/sessions/{sessionKey}/forms/{formInstanceId}:
    get:
      tags: [ACC45]
      summary: Retrieve a previously saved form's rendered view/definition
      parameters:
        - $ref: '#/components/parameters/sessionKey'
        - $ref: '#/components/parameters/formInstanceId'
      responses:
        '200':
          description: Saved form view/definition
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FormData' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [ACC45]
      summary: Save a completed form (render to DMS, then persist ACC45 definition/diagnosis/referral — HISO-BR-12 sequencing)
      parameters:
        - $ref: '#/components/parameters/sessionKey'
        - $ref: '#/components/parameters/formInstanceId'
        - $ref: '#/components/parameters/idempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FormSaveInput' }
      responses:
        '200':
          description: Form saved
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FormData' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /acc45/sessions/{sessionKey}/actions:
    post:
      tags: [ACC45]
      summary: >
        Dispatch a form action: save, addTask, addInvoice, launchForm. The latter
        two were unimplemented no-op stubs in legacy HISO; implemented for real
        per stakeholder decision 2026-07-19 (Section 8 Decision 4 of the contract doc).
      parameters:
        - $ref: '#/components/parameters/sessionKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [actionId]
              properties:
                actionId: { type: string, enum: [save, addTask, addInvoice, launchForm] }
                actionContainer: { type: object }
      responses:
        '200':
          description: Action processed
          content:
            application/json:
              schema:
                type: object
                properties:
                  processed: { type: boolean }
...
  /patients/{patientId}/invoices:
    post:
      tags: [Billing]
      summary: Save an invoice/service record (idempotent on duplicate; requires billing:write scope)
      security:
        - bearerAuth: [billing:write]
      parameters:
        - $ref: '#/components/parameters/patientId'
        - $ref: '#/components/parameters/idempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvoiceInput'
      responses:
        '200':
          description: Duplicate submission — existing invoice returned, not an error (FR-IDEM-01)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Invoice' }
        '201':
          description: Invoice created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Invoice' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /admin/practices:
    post:
      tags: [Admin]
      summary: Register a new practice in the tenant registry (FR-TEN-01, ADR-001)
      security:
        - bearerAuth: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PracticeInput'
      responses:
        '201':
          description: Practice registered
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Practice' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /admin/practices/{practiceId}:
    get:
      tags: [Admin]
      summary: Retrieve a practice's registry entry
      security:
        - bearerAuth: [admin]
      parameters:
        - name: practiceId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Practice registry entry
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Practice' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Admin]
      summary: Update a practice's routing/config entry
      security:
        - bearerAuth: [admin]
      parameters:
        - name: practiceId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PracticeInput'
      responses:
        '200':
          description: Practice updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Practice' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Resource-scoped token per ADR-003 (patientId + encounterId + practiceId +
        structurally-determined originScope). Issued via POST /auth/token, or
        internally minted by the HISO SOAP edge adapter from an existing
        SessionGUID (ADR-004/ADR-007).
  parameters:
    patientId:
      name: patientId
      in: path
      required: true
      schema: { type: integer }
    encounterId:
      name: encounterId
      in: path
      required: true
      schema: { type: integer }
...
    sessionKey:
      name: sessionKey
      in: path
      required: true
      schema: { type: string, format: uuid }
    formInstanceId:
      name: formInstanceId
      in: path
      required: true
      schema: { type: string }
    sinceDate:
      name: sinceDate
      in: query
      schema: { type: string, format: date }
    untilDate:
      name: untilDate
      in: query
      schema: { type: string, format: date }
    sortOrder:
      name: sortOrder
      in: query
      schema: { type: string, enum: [asc, desc], default: desc }
    idempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema: { type: string }
      description: See Section 12 of the contract document. Optional; falls back to natural-key duplicate detection if omitted.
  responses:
    BadRequest:
      description: Validation failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Missing/invalid/expired credential
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Valid credential, insufficient scope
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: Genuine conflict outside the documented idempotency path
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      description: Data-minimized error shape — never contains PHI, stack traces, or internal exception detail (SRS §12.14).
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string, description: Generic, static, non-PHI message. }
        traceId: { type: string }
        errors:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }
      required: [title, status]
...
    ClinicalNote:
      type: object
      properties:
        noteId: { type: string }
        patientId: { type: integer }
        encounterId: { type: integer }
        author: { type: string }
        createdAt: { type: string, format: date-time }
        content: { type: string }
    ClinicalNoteInput:
      type: object
      required: [content]
      properties:
        content: { type: string }
    Condition:
      type: object
      properties:
        conditionId: { type: string }
        status: { type: string, enum: [created, existing] }
        diagnosisCode: { type: string }
        description: { type: string }
        isLongTerm: { type: boolean }
        sideCode: { type: string, nullable: true }
        sideDescription: { type: string, nullable: true }
    ConditionInput:
      type: object
      required: [diagnosisCode]
      properties:
        diagnosisCode: { type: string }
        description: { type: string }
        isLongTerm: { type: boolean, default: false }
        sideCode: { type: string, nullable: true }
        sideDescription: { type: string, nullable: true }
...
    SessionContext:
      type: object
      properties:
        providerId: { type: string }
        patientId: { type: integer }
        appointmentId: { type: string }
        practiceId: { type: string }
        referenceId: { type: string, nullable: true }
        practiceLocationId: { type: string, nullable: true }
        practiceEdi: { type: string, nullable: true }
    FormData:
      type: object
      properties:
        formInstanceId: { type: string }
        viewType: { type: string }
        view: { type: string }
        dataContainer: { type: object, additionalProperties: true }
        resumePath: { type: string, nullable: true }
    FormSaveInput:
      type: object
      required: [dataContainer, completed]
      properties:
        dataContainer: { type: object, additionalProperties: true }
        view: { type: string }
        viewType: { type: string }
        completed: { type: boolean }
        continueSession: { type: boolean }
...
    InvoiceInput:
      type: object
      required: [serviceCode, amountInclGst]
      properties:
        serviceCode: { type: string }
        serviceName: { type: string }
        amountInclGst: { type: number }
        payee: { type: string }
        serviceProvider: { type: string }
        serviceDate: { type: string, format: date }
        pegasusReference: { type: string, nullable: true }
        claimShortCode: { type: string, nullable: true }
    Invoice:
      allOf:
        - $ref: '#/components/schemas/InvoiceInput'
        - type: object
          properties:
            invoiceId: { type: string }
            status: { type: string, enum: [created, existing] }
    PracticeInput:
      type: object
      required: [name, databaseServerId]
      properties:
        name: { type: string }
        phoCode: { type: string, nullable: true }
        databaseServerId: { type: string }
    Practice:
      allOf:
        - $ref: '#/components/schemas/PracticeInput'
        - type: object
          properties:
            practiceId: { type: string }